requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
certifi==2025.6.15
click==8.2.1
coloredlogs==15.0.1
fastapi==0.115.14
flatbuffers==25.2.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
kafka-python==2.2.14
mpmath==1.3.0
numpy==2.3.1
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
packaging==25.0
paho-mqtt==2.1.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
python-snappy==0.7.3
PyYAML==6.0.2
rtmlib==0.0.13
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
